+++ src/wrapped/wrappedmpg123.c
@@ -87,10 +87,10 @@ static void* find_r_read_Fct(void* fct)
 }
 // r_lseek ...
 #define GO(A)   \
-static uintptr_t my_r_lseek_fct_##A = 0;                                 \
-static int64_t my_r_lseek_##A(void* a, int64_t b, int n)             \
-{                                                                                       \
-    return (int64_t)RunFunction64(my_context, my_r_lseek_fct_##A, 4, a, b&0xffffffff, (b>>32)&0xffffffff, n); \
+static uintptr_t my_r_lseek_fct_##A = 0;                                                                                            \
+static int64_t my_r_lseek_##A(void* a, int64_t b, int n)                                                                            \
+{                                                                                                                                   \
+    return (int64_t)RunFunction64(my_context, my_r_lseek_fct_##A, 4, a, (uint32_t)(b&0xffffffff), (uint32_t)(b>>32)&0xffffffff, n); \
 }
 SUPER()
 #undef GO
